# openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "electrostore",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "description": "URL de base personnalisable",
      "variables": {
        "baseUrl": {
          "default": "https://exemple.com",
          "description": "URL de base personnalisée",
          "examples": [
            "https://api.exemple.com",
            "https://staging-api.exemple.com",
            "http://localhost:8080"
          ]
        }
      }
    }
  ],
  "host": "{baseUrl}",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/store": {
      "get": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of items",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadStoreDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadStoreDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadStoreDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Stores"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/{id_store}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/tag/{id_tag}/store/{id_store}": {
      "get": {
        "tags": [
          "TagStore"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TagStore"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TagStore"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateStoreDto": {
        "required": [
          "mqtt_name_store",
          "nom_store",
          "xlength_store",
          "ylength_store"
        ],
        "type": "object",
        "properties": {
          "nom_store": {
            "minLength": 1,
            "type": "string"
          },
          "xlength_store": {
            "type": "integer",
            "format": "int32"
          },
          "ylength_store": {
            "type": "integer",
            "format": "int32"
          },
          "mqtt_name_store": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ReadStoreDto": {
        "type": "object",
        "properties": {
          "id_store": {
            "type": "integer",
            "format": "int32"
          },
          "nom_store": {
            "type": "string",
            "nullable": true
          },
          "xlength_store": {
            "type": "integer",
            "format": "int32"
          },
          "ylength_store": {
            "type": "integer",
            "format": "int32"
          },
          "mqtt_name_store": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReadStoreTagDto": {
        "type": "object",
        "properties": {
          "id_store": {
            "type": "integer",
            "format": "int32"
          },
          "id_tag": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateStoreDto": {
        "type": "object",
        "properties": {
          "nom_store": {
            "type": "string",
            "nullable": true
          },
          "xlength_store": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ylength_store": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "mqtt_name_store": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [
        
      ]
    }
  ]
}